Prevent autocomplete on suggestions that don't match
The behaviour of the ngTagsInput directive is somewhat counter-intuitive -
typing part of a suggested email will appear as though it matches but pressing
enter will not input the suggestion, but the current typed value.

Disabling add on enter will still allow enter to select the selection, but
prevents selection of the partially typed email/name.