Don't check for linting errors if code check is disabled
There are a couple of reasons for this:

1. Some linting errors are returned from the server after a compile is run,
replacing client-side linting errors. If code check is disabled this does not
happen, and therefore linting errors persist until the next compile. This makes
it appear as though autocompile is not running

2. It is likely that if code check is disabled, the user is deliberately
ignoring linting errors and therefore the linting check is pointless