Remove unused contentful dependency
This package is unused, and was causing problems with installing babel. Because
it dependended on an older version of babel, and was interpreted by npm first
(as it is a regular dep), the older version of babel would be "hoisted" to the
top level of node_modules

Then when latest versions of babel are installed they would be interpreted second
and thus would nest their dependencies within their own node_modules, filling the
disk with unnecessary copies of core-js.

Removing the unnecessary dep solves these problems