Prevent autocompile loop
If an autocompile hits a rate limit, it would get into a bad state where
it would continuely loop making compile requests. This is because the
compile response resolving would disable auto compile (because the rate
limit was hit) but also trigger calculation of uncompiled changes, which
would not check if autocompile was disabled. The fix is to just check if
autocompile is disabled.