Update sharelatex.conf to fix IPv6 issues reported in #1146
This changes the resolution to the underlying sharelatex service to use "localhost" instead of the IPv4 address 127.0.0.1. This is because node.js will bind to whatever DNS record is returned for the binding address given, which is "localhost" in the current configuration, which leads to a binding to ::1 instead of 127.0.0.1. Either we change in the app.js for sharelatex to always force IPv4, or we just use whatever localhost binds to.